refactor: separate logic for Sensor Package into its own source files
