+++ Arduino/platformio.ini
@@ -27,6 +27,7 @@ build_flags =
     -Wunreachable-code
     -Wduplicated-cond
     -Wduplicated-branches
+    -Iinclude
 lib_deps =
 	adafruit/Adafruit Unified Sensor@^1.1.15
 	adafruit/DHT sensor library@^1.4.6
@@ -46,6 +47,7 @@ build_flags =
     -Wunreachable-code
     -Wduplicated-cond
     -Wduplicated-branches
+    -Iinclude
 lib_deps =
 	adafruit/Adafruit Unified Sensor@^1.1.15
 	adafruit/DHT sensor library@^1.4.6
